release(social-platform): integrate develop social and reliability updates

+++ CloudM.Application/Services/CommentReactServices/CommentReactService.cs
@@ -6,11 +6,13 @@ using CloudM.Domain.Entities;
 using CloudM.Domain.Enums;
 using CloudM.Infrastructure.Models;
 using CloudM.Infrastructure.Repositories.Accounts;
+using CloudM.Infrastructure.Repositories.AccountBlocks;
 using CloudM.Infrastructure.Repositories.CommentReacts;
 using CloudM.Infrastructure.Repositories.Comments;
 using CloudM.Infrastructure.Repositories.Posts;
 using CloudM.Infrastructure.Repositories.Follows;
 using CloudM.Infrastructure.Repositories.UnitOfWork;
+using CloudM.Application.Services.NotificationServices;
 using CloudM.Application.Services.RealtimeServices;
 using System;
 using System.Collections.Generic;
@@ -29,11 +31,14 @@ namespace CloudM.Application.Services.CommentReactServices
         private readonly IFollowRepository _followRepository;
         private readonly IUnitOfWork _unitOfWork;
         private readonly IMapper _mapper;
+        private readonly INotificationService _notificationService;
         private readonly IRealtimeService _realtimeService;
+        private readonly IAccountBlockRepository _accountBlockRepository;
 
         public CommentReactService(ICommentRepository commentRepository, ICommentReactRepository commentReactRepository, IPostRepository postRepository,
             IAccountRepository accountRepository, IFollowRepository followRepository, IMapper mapper, 
-            IRealtimeService realtimeService, IUnitOfWork unitOfWork)
+            INotificationService notificationService, IRealtimeService realtimeService, IUnitOfWork unitOfWork,
+            IAccountBlockRepository? accountBlockRepository = null)
         {
             _commentRepository = commentRepository;
             _commentReactRepository = commentReactRepository;
@@ -41,8 +46,31 @@ namespace CloudM.Application.Services.CommentReactServices
             _accountRepository = accountRepository;
             _followRepository = followRepository;
             _mapper = mapper;
+            _notificationService = notificationService;
             _realtimeService = realtimeService;
             _unitOfWork = unitOfWork;
+            _accountBlockRepository = accountBlockRepository ?? NullAccountBlockRepository.Instance;
+        }
+        public CommentReactService(
+            ICommentRepository commentRepository,
+            ICommentReactRepository commentReactRepository,
+            IPostRepository postRepository,
+            IAccountRepository accountRepository,
+            IFollowRepository followRepository,
+            IMapper mapper,
+            IRealtimeService realtimeService,
+            IUnitOfWork unitOfWork)
+            : this(
+                commentRepository,
+                commentReactRepository,
+                postRepository,
+                accountRepository,
+                followRepository,
+                mapper,
+                NullNotificationService.Instance,
+                realtimeService,
+                unitOfWork)
+        {
         }
         public async Task<ReactToggleResponse> ToggleReactOnComment(Guid commentId, Guid accountId)
         {
@@ -58,14 +86,24 @@ namespace CloudM.Application.Services.CommentReactServices
                 throw new BadRequestException($"Post with ID {comment.PostId} not found.");
             }
 
+            if (await _accountBlockRepository.IsBlockedEitherWayAsync(accountId, comment.AccountId) ||
+                await _accountBlockRepository.IsBlockedEitherWayAsync(accountId, post.AccountId))
+                throw new BadRequestException("This content is no longer available.");
+
             await ValidatePostPrivacyAsync(post, accountId, "react to comments on");
 
+            var nowUtc = DateTime.UtcNow;
             var existingReact = await _commentReactRepository.GetUserReactOnCommentAsync(commentId, accountId);
             var isReactedByCurrentUser = false;
             if (existingReact != null)
             {
                 await _commentReactRepository.RemoveCommentReact(existingReact);
                 isReactedByCurrentUser = false;
+                await EnqueueCommentReactNotificationAsync(
+                    comment,
+                    accountId,
+                    NotificationAggregateActionEnum.Deactivate,
+                    nowUtc);
             }
             else
             {
@@ -74,10 +112,15 @@ namespace CloudM.Application.Services.CommentReactServices
                     CommentId = commentId,
                     AccountId = accountId,
                     ReactType = ReactEnum.Love,
-                    CreatedAt = DateTime.UtcNow
+                    CreatedAt = nowUtc
                 };
                 await _commentReactRepository.AddCommentReact(newReact);
                 isReactedByCurrentUser = true;
+                await EnqueueCommentReactNotificationAsync(
+                    comment,
+                    accountId,
+                    NotificationAggregateActionEnum.Upsert,
+                    nowUtc);
             }
 
             // Must commit before counting
@@ -94,6 +137,37 @@ namespace CloudM.Application.Services.CommentReactServices
                 IsReactedByCurrentUser = isReactedByCurrentUser
             };
         }
+        private async Task EnqueueCommentReactNotificationAsync(
+            Comment comment,
+            Guid actorId,
+            NotificationAggregateActionEnum action,
+            DateTime occurredAt)
+        {
+            if (comment.AccountId == Guid.Empty || actorId == Guid.Empty || comment.AccountId == actorId)
+            {
+                return;
+            }
+
+            var isReply = comment.ParentCommentId.HasValue;
+            await _notificationService.EnqueueAggregateEventAsync(new NotificationAggregateEvent
+            {
+                RecipientId = comment.AccountId,
+                Action = action,
+                Type = isReply ? NotificationTypeEnum.ReplyReact : NotificationTypeEnum.CommentReact,
+                AggregateKey = isReply
+                    ? NotificationAggregateKeys.ReplyReact(comment.CommentId)
+                    : NotificationAggregateKeys.CommentReact(comment.CommentId),
+                SourceType = isReply
+                    ? NotificationSourceTypeEnum.ReplyReact
+                    : NotificationSourceTypeEnum.CommentReact,
+                SourceId = actorId,
+                ActorId = actorId,
+                TargetKind = NotificationTargetKindEnum.Post,
+                TargetId = comment.PostId,
+                KeepWhenEmpty = false,
+                OccurredAt = occurredAt
+            });
+        }
         public async Task<PagedResponse<AccountReactListModel>> GetAccountsReactOnCommentPaged(Guid commentId, Guid? currentId, int page, int pageSize)
         {
             var comment = await _commentRepository.GetCommentById(commentId);
@@ -107,6 +181,11 @@ namespace CloudM.Application.Services.CommentReactServices
             {
                 throw new BadRequestException($"Post with ID {comment.PostId} not found.");
             }
+
+            if (currentId.HasValue &&
+                (await _accountBlockRepository.IsBlockedEitherWayAsync(currentId.Value, comment.AccountId) ||
+                 await _accountBlockRepository.IsBlockedEitherWayAsync(currentId.Value, post.AccountId)))
+                throw new BadRequestException("This content is no longer available.");
 
             if (currentId.HasValue)
             {
